Base the line number off of the lines list length
Rather than having a separate variable that keeps track of the number of lines,
use a property that returns the length of the list of lines. This also allows
us to set line numbers as strings, which is allowed through the Avatax API.

This also fixes a bug, where if functions using the line_num variable are called
before the `with_line` function a TypeError is raised. This was due to the
line_num variable not being instantiated until after the first time the
`with_line` function is called.